Rename user defined data writer file
